multiple fixes
fixed the search button on the index & update page to be functional.
Also added sticky-top for the nav so as to allow easy access to the search bar.
Added a Home button on the update.html to allow for easy navigation to the home page. This is in addition to the cancel and close button, if the users changes mind.